+++ crates/test_mode_bot/src/utils/location.rs
@@ -32,6 +32,6 @@ impl LocationConfigUtils for LocationConfig {
         let random_latitude = self.latitude_bottom_right + (y_len * y_precent);
         let random_longitude = self.longitude_bottom_right + (x_len * x_precent);
 
-        Location::new(random_latitude, random_longitude)
+        Location::new(random_latitude as f32, random_longitude as f32)
     }
 }
